version 2
Se realizan modificaciones, correccion de errores, se realizan pruebas, y se deja operativo el codigo

+++ SAFE-Rescue_Backend_API-Ciudadano/src/main/java/com/SAFE_Rescue/API_Ciudadano/service/CiudadanoService.java
@@ -126,6 +126,9 @@ public class CiudadanoService {
 
     public void delete(long id){
         try {
+            if (!ciudadanoRepository.existsById(id)) {
+                throw new NoSuchElementException("Ciudadano no encontrado");
+            }
             ciudadanoRepository.deleteById(id);
 
         } catch (Exception e) {
